third_party/WebKit/LayoutTests/external/wpt/web-nfc/nfc_push.https-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = WebNFC is not supported.
PASS Test that promise is rejected with TypeError if NFCMessage is invalid.
PASS 'Test that promise is rejected with SyntaxError if NFCMessage contains invalid records.
PASS nfc.push should fail with TypeError when invalid timeout is provided.
PASS nfc.push should fail with TypeError when invalid negative timeout value is provided.
FAIL nfc.push should fail with TimeoutError when timer expires. assert_throws: function "function() { throw e }" threw object "NotSupportedError: NFC operation not supported." that is not a DOMException TimeoutError: property "code" is equal to 9, expected 23
FAIL nfc.cancelPush should reject pending promise with AbortError. assert_throws: function "function() { throw e }" threw object "NotSupportedError: WebNFC is not supported." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
PASS Reject promise with NotSupportedError if NFC message size exceeds 32KB.
FAIL Reject promise with SyntaxError if WebNFC Id cannot be created from provided URL. assert_throws: function "function() { throw e }" threw object "NotSupportedError: WebNFC is not supported." that is not a DOMException SyntaxError: property "code" is equal to 9, expected 12
FAIL Reject promise with SyntaxError if 'json' record cannot be serialized. assert_throws: function "function() { throw e }" threw object "NotSupportedError: WebNFC is not supported." that is not a DOMException SyntaxError: property "code" is equal to 9, expected 12
PASS nfc.push should fail with TypeError when invalid target value is provided.
PASS nfc.cancelPush should fail with TypeError when invalid id value is provided.
PASS Test that WebNFC API is not accessible from iframe context.
Harness: the test ran to completion.
